+++ src/basemath/trans1.c
@@ -2283,7 +2283,7 @@ atanhQ_split(ulong u, ulong v, long prec)
   struct abpq_res R;
   struct abpq A;
   /* satisfies (2n+1) (v/u)^2n > 2^bitprec */
-  nmax = bit_accuracy(prec) / (2*log2(d));
+  nmax = (long)(bit_accuracy(prec) / (2*log2(d)));
   abpq_init(&A, nmax);
   A.a[0] = A.b[0] = gen_1;
   A.p[0] = utoipos(u);
